Rename cluster.snapshot to cluster.temporal
