+++ doc/api/etrago.cluster.rst
@@ -18,10 +18,10 @@ etrago.cluster.gas module
     :undoc-members:
     :show-inheritance:
 
-etrago.cluster.snapshot module
+etrago.cluster.temporal module
 --------------------------------
 
-.. automodule:: etrago.cluster.snapshot
+.. automodule:: etrago.cluster.temporal
     :members:
     :undoc-members:
     :show-inheritance:
